Revert "#3370 move command inside stimmzettelerfassung"
This reverts commit 7499e15758d5ceb2bc4440ccdecfb53cc103d73a.
