+++ wls-gui-wahllokalsystem/src/types/dse/command/CommandHandler.ts
